Begin refactoring KnitWindow, fix undoing bug
I didn't want to have to redo everything for a window that would work
with Motifs instead of Designs, so I'm refactoring KnitWindow. This is
working on #58

Also fixed an undoing bug that happened when you tried to put down a
motif but there was an overlap or it would go off the design, and that
improper action would get added to the design program builder anyway.
Fixes #59