Got a harebrained idea to use scala to write one source of truth
since i'm moving between TypeScript and Python, and I'd like to keep the
option of having the Python bindings while not having to update the code
manually, I thought it would be a good idea to try to use scala to
generate the Python and TypeScript code.

I actually think this might work fairly well for generic stuff.